Move patch and conditions packages into v1a1 webconsolerequest controller
These packages operation on the v1a1 types only. Since the name of
the WebConsoleRequest CRD change in v1a2 version conversion is
not possible the v1a1 controller is the only place they are still
used.
