+++ controllers/virtualmachinewebconsolerequest/v1alpha1/conditions/merge_strategies.go
