+++ controllers/virtualmachinewebconsolerequest/v1alpha1/conditions/patch.go
