+++ controllers/virtualmachinewebconsolerequest/v1alpha1/conditions/utils.go
