+++ controllers/virtualmachinewebconsolerequest/v1alpha1/patch/patch_test.go
@@ -32,8 +32,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	vmopv1 "github.com/vmware-tanzu/vm-operator/api/v1alpha1"
-
-	"github.com/vmware-tanzu/vm-operator/pkg/conditions"
+	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachinewebconsolerequest/v1alpha1/conditions"
 	"github.com/vmware-tanzu/vm-operator/test/builder"
 )
