+++ controllers/virtualmachinewebconsolerequest/v1alpha1/patch/utils_test.go
